Fixed the postgres constraint added for enum.
MySQL is weird. When a value not specified in enum is added, instead of
giving an error, it inserts empty string and happily ignores the given
value.

When we are migrating a MySQL database with such data, we should make
sure the existing data loads without any errors.

Added '' as one of the allowed values along with the values specified in
enum to the CHECK constraint to handle this.